+++ src/icemac/addressbook/browser/wizard.py
@@ -8,6 +8,8 @@ import z3c.wizard.wizard
 class Step(z3c.wizard.step.Step):
     """Customized wizard step which (easier to use)."""
 
+    title = None
+
     @property
     def fields(self):
         return z3c.form.field.Fields(self.interface)
@@ -15,3 +17,5 @@ class Step(z3c.wizard.step.Step):
 
 class Wizard(z3c.wizard.wizard.Wizard):
     """Base (customizable) wizard."""
+
+    title = None